refactor(web): atomize filter routes into forms/handlers/responses (#657)
filter.py mixed form parsing, orchestration, and redirect/response mapping
in one module. Split onto the established route-layering convention:

- src/web/filter/forms.py — parse_snapshot/parse_pks/parse_confirm_pairs +
  HARD_DELETE_ALL_CONFIRM_PHRASE.
- src/web/filter/handlers.py — orchestration (analyzer, deletion service,
  dev-mode gate, hard-delete-all set comparison), returns domain DTOs.
- src/web/filter/responses.py — FilterRedirect/FilterTemplate + filter_response
  mapper, reusing the shared flash_redirect helper.
- src/web/routes/filter.py — thin router delegating to handlers + mapper.

Route paths, redirect codes/params, and template contexts are unchanged.
Test patch targets repointed to the new handlers module (deps/ChannelAnalyzer
moved there); assertions unchanged. Part of epic #402.
