+++ tests/routes/test_filter_routes.py
@@ -51,7 +51,7 @@ async def test_purge_selected_success(route_client, db):
     """Test purge selected channels."""
     pk = await _add_filtered_channel(db, channel_id=400, title="To Purge")
 
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         mock_svc.return_value.purge_channels_by_pks = AsyncMock()
         resp = await route_client.post(
             "/channels/filter/purge-selected",
@@ -67,7 +67,7 @@ async def test_purge_selected_removes_messages(route_client, db):
     """Test purge selected removes messages from DB."""
     pk = await _add_filtered_channel(db, channel_id=500, title="Purge Messages")
 
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         mock_svc.return_value.purge_channels_by_pks = AsyncMock()
         resp = await route_client.post(
             "/channels/filter/purge-selected",
@@ -83,7 +83,7 @@ async def test_purge_selected_removes_messages(route_client, db):
 @pytest.mark.anyio
 async def test_purge_all_no_filtered(route_client):
     """Test purge all with no filtered channels."""
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         from src.services.filter_deletion_service import PurgeResult
         mock_svc.return_value.purge_all_filtered = AsyncMock(
             return_value=PurgeResult(purged_count=0)
@@ -98,7 +98,7 @@ async def test_purge_all_success(route_client, db):
     """Test purge all filtered channels."""
     await _add_filtered_channel(db, channel_id=600, title="Purge All")
 
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         from src.services.filter_deletion_service import PurgeResult
         mock_svc.return_value.purge_all_filtered = AsyncMock(
             return_value=PurgeResult(purged_count=1)
@@ -140,7 +140,7 @@ async def test_hard_delete_selected_partial_failure_reports_error(route_client, 
     pk1 = await _add_filtered_channel(db, channel_id=810, title="Sel OK")
     pk2 = await _add_filtered_channel(db, channel_id=811, title="Sel Skip")
     await _enable_dev_mode(db)
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
             return_value=PurgeResult(purged_count=1, skipped_count=1)
         )
@@ -163,7 +163,7 @@ async def test_hard_delete_success(route_client, db):
     pk = await _add_filtered_channel(db, channel_id=800, title="Hard Delete OK")
     await _enable_dev_mode(db)
 
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         from src.services.filter_deletion_service import PurgeResult
         mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
             return_value=PurgeResult(purged_count=1)
@@ -196,7 +196,7 @@ async def test_hard_delete_all_rejects_without_confirm_phrase(route_client, db):
     when dev mode is on — a direct POST without it must NOT delete."""
     pk = await _add_filtered_channel(db, channel_id=901, title="No Confirm")
     await _enable_dev_mode(db)
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
             return_value=PurgeResult(purged_count=1)
         )
@@ -217,7 +217,7 @@ async def test_hard_delete_all_rejects_wrong_confirm_phrase(route_client, db):
     """A non-matching confirm value must be rejected (no fuzzy match)."""
     pk = await _add_filtered_channel(db, channel_id=902, title="Wrong Confirm")
     await _enable_dev_mode(db)
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
             return_value=PurgeResult(purged_count=1)
         )
@@ -236,7 +236,7 @@ async def test_hard_delete_all_rejects_missing_confirm_pks(route_client, db):
     """confirm_pks is required: a request without it cannot reach the delete."""
     await _add_filtered_channel(db, channel_id=903, title="No PKs")
     await _enable_dev_mode(db)
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
             return_value=PurgeResult(purged_count=1)
         )
@@ -264,7 +264,7 @@ async def test_hard_delete_all_rejects_malformed_confirm_pks(route_client, db):
         ":1",            # empty pk
     ]
     for raw in cases:
-        with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+        with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
             mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
                 return_value=PurgeResult(purged_count=1)
             )
@@ -291,7 +291,7 @@ async def test_hard_delete_all_rejects_duplicate_confirm_pks(route_client, db):
         f"{pk}:970,99:970",     # duplicate chid, different pk
     ]
     for raw in cases:
-        with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+        with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
             mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
                 return_value=PurgeResult(purged_count=1)
             )
@@ -319,7 +319,7 @@ async def test_hard_delete_all_rejects_pk_with_wrong_channel_id(route_client, db
     """
     pk = await _add_filtered_channel(db, channel_id=975, title="Stable")
     await _enable_dev_mode(db)
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
             return_value=PurgeResult(purged_count=1)
         )
@@ -341,7 +341,7 @@ async def test_hard_delete_all_rejects_set_mismatch(route_client, db):
     """Stale page snapshots a different filtered set — bounce."""
     pk = await _add_filtered_channel(db, channel_id=920, title="OnlyMe")
     await _enable_dev_mode(db)
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
             return_value=PurgeResult(purged_count=1)
         )
@@ -372,7 +372,7 @@ async def test_hard_delete_all_rejects_same_count_stale_swap(route_client, db):
     await db.set_channel_filtered(a_pk, False)
     b_pk = await _add_filtered_channel(db, channel_id=941, title="ChB")
     await _enable_dev_mode(db)
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
             return_value=PurgeResult(purged_count=1)
         )
@@ -397,7 +397,7 @@ async def test_hard_delete_all_success_with_matching_snapshot(route_client, db):
     the delete, and the service is called with exactly the confirmed PKs."""
     pk = await _add_filtered_channel(db, channel_id=950, title="Will Delete")
     await _enable_dev_mode(db)
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
             return_value=PurgeResult(purged_count=1)
         )
@@ -424,7 +424,7 @@ async def test_hard_delete_all_rejects_empty_confirm(route_client, db):
     await _enable_dev_mode(db)
     cases = ["", "   ", "\t", "\n"]
     for raw in cases:
-        with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+        with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
             mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
                 return_value=PurgeResult(purged_count=1)
             )
@@ -450,7 +450,7 @@ async def test_hard_delete_all_partial_failure_reports_error(route_client, db):
     pk1 = await _add_filtered_channel(db, channel_id=981, title="ChDelOK")
     pk2 = await _add_filtered_channel(db, channel_id=982, title="ChDelFail")
     await _enable_dev_mode(db)
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         # Service deleted pk1 but skipped pk2.
         mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
             return_value=PurgeResult(purged_count=1, skipped_count=1)
@@ -478,7 +478,7 @@ async def test_hard_delete_all_purge_count_mismatch_reports_error(route_client, 
     the route still surfaces it as an error rather than a silent partial."""
     pk = await _add_filtered_channel(db, channel_id=983, title="Mismatch")
     await _enable_dev_mode(db)
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         # purged=0, skipped=0 — count mismatches expected (1).
         mock_svc.return_value.hard_delete_channels_by_pks = AsyncMock(
             return_value=PurgeResult(purged_count=0, skipped_count=0)
@@ -510,7 +510,7 @@ async def test_hard_delete_all_no_filtered_channels(route_client, db):
 @pytest.mark.anyio
 async def test_analyze_redirects(route_client):
     """Test analyze channels redirects."""
-    with patch("src.web.routes.filter.ChannelAnalyzer") as mock_analyzer:
+    with patch("src.web.filter.handlers.ChannelAnalyzer") as mock_analyzer:
         from src.filters.models import FilterReport
         mock_instance = mock_analyzer.return_value
         mock_instance.analyze_all = AsyncMock(
@@ -525,14 +525,14 @@ async def test_analyze_redirects(route_client):
 @pytest.mark.anyio
 async def test_analyze_ignores_with_stats_query(route_client):
     """Test analyze route no longer runs stats collection inline."""
-    with patch("src.web.routes.filter.ChannelAnalyzer") as mock_analyzer:
+    with patch("src.web.filter.handlers.ChannelAnalyzer") as mock_analyzer:
         from src.filters.models import FilterReport
         mock_instance = mock_analyzer.return_value
         mock_instance.analyze_all = AsyncMock(
             return_value=FilterReport(results=[], total_channels=0, filtered_count=0)
         )
         mock_instance.apply_filters = AsyncMock(return_value=0)
-        with patch("src.web.routes.filter.deps.collection_service") as mock_collection:
+        with patch("src.web.filter.handlers.deps.collection_service") as mock_collection:
             resp = await route_client.post("/channels/filter/analyze?with_stats=1", follow_redirects=False)
 
         assert resp.status_code == 303
@@ -586,7 +586,7 @@ async def test_apply_missing_snapshot(route_client):
 @pytest.mark.anyio
 async def test_apply_with_snapshot(route_client):
     """Test apply filters with snapshot."""
-    with patch("src.web.routes.filter.ChannelAnalyzer") as mock_analyzer:
+    with patch("src.web.filter.handlers.ChannelAnalyzer") as mock_analyzer:
         mock_instance = mock_analyzer.return_value
         mock_instance.apply_filters = AsyncMock(return_value=1)
         resp = await route_client.post(
@@ -601,7 +601,7 @@ async def test_apply_with_snapshot(route_client):
 @pytest.mark.anyio
 async def test_precheck_redirects(route_client):
     """Test precheck subscriber ratio redirects."""
-    with patch("src.web.routes.filter.ChannelAnalyzer") as mock_analyzer:
+    with patch("src.web.filter.handlers.ChannelAnalyzer") as mock_analyzer:
         mock_instance = mock_analyzer.return_value
         mock_instance.precheck_subscriber_ratio = AsyncMock(return_value=5)
         resp = await route_client.post("/channels/filter/precheck", follow_redirects=False)
@@ -612,7 +612,7 @@ async def test_precheck_redirects(route_client):
 @pytest.mark.anyio
 async def test_reset_redirects(route_client):
     """Test reset filters redirects."""
-    with patch("src.web.routes.filter.ChannelAnalyzer") as mock_analyzer:
+    with patch("src.web.filter.handlers.ChannelAnalyzer") as mock_analyzer:
         mock_instance = mock_analyzer.return_value
         mock_instance.reset_filters = AsyncMock()
         resp = await route_client.post("/channels/filter/reset", follow_redirects=False)
@@ -666,7 +666,7 @@ async def test_filter_toggle_success(route_client, db):
 @pytest.mark.anyio
 async def test_parse_snapshot_valid(route_client, db):
     """Test apply filters with valid snapshot parsing."""
-    with patch("src.web.routes.filter.ChannelAnalyzer") as mock_analyzer:
+    with patch("src.web.filter.handlers.ChannelAnalyzer") as mock_analyzer:
         mock_instance = mock_analyzer.return_value
         mock_instance.apply_filters = AsyncMock(return_value=2)
         resp = await route_client.post(
@@ -684,7 +684,7 @@ async def test_parse_snapshot_valid(route_client, db):
 @pytest.mark.anyio
 async def test_parse_snapshot_dedupes_by_channel_id(route_client, db):
     """Test snapshot parsing dedupes by channel_id."""
-    with patch("src.web.routes.filter.ChannelAnalyzer") as mock_analyzer:
+    with patch("src.web.filter.handlers.ChannelAnalyzer") as mock_analyzer:
         mock_instance = mock_analyzer.return_value
         mock_instance.apply_filters = AsyncMock(return_value=1)
         resp = await route_client.post(
@@ -701,7 +701,7 @@ async def test_parse_snapshot_dedupes_by_channel_id(route_client, db):
 @pytest.mark.anyio
 async def test_parse_snapshot_invalid_channel_id(route_client, db):
     """Test snapshot parsing with invalid channel_id."""
-    with patch("src.web.routes.filter.ChannelAnalyzer") as mock_analyzer:
+    with patch("src.web.filter.handlers.ChannelAnalyzer") as mock_analyzer:
         mock_instance = mock_analyzer.return_value
         mock_instance.apply_filters = AsyncMock(return_value=0)
         resp = await route_client.post(
@@ -718,7 +718,7 @@ async def test_parse_snapshot_invalid_channel_id(route_client, db):
 @pytest.mark.anyio
 async def test_parse_snapshot_no_separator(route_client, db):
     """Test snapshot parsing without separator."""
-    with patch("src.web.routes.filter.ChannelAnalyzer") as mock_analyzer:
+    with patch("src.web.filter.handlers.ChannelAnalyzer") as mock_analyzer:
         mock_instance = mock_analyzer.return_value
         mock_instance.apply_filters = AsyncMock(return_value=0)
         resp = await route_client.post(
@@ -735,7 +735,7 @@ async def test_parse_snapshot_no_separator(route_client, db):
 @pytest.mark.anyio
 async def test_parse_snapshot_invalid_flag(route_client, db):
     """Test snapshot parsing with invalid flag."""
-    with patch("src.web.routes.filter.ChannelAnalyzer") as mock_analyzer:
+    with patch("src.web.filter.handlers.ChannelAnalyzer") as mock_analyzer:
         mock_instance = mock_analyzer.return_value
         mock_instance.apply_filters = AsyncMock(return_value=0)
         resp = await route_client.post(
@@ -755,8 +755,8 @@ async def test_analyze_with_auto_delete(route_client, db):
     await _add_filtered_channel(db, channel_id=3100, title="Auto Delete")
     await db.set_setting("auto_delete_filtered", "1")
 
-    with patch("src.web.routes.filter.ChannelAnalyzer") as mock_analyzer, patch(
-        "src.web.routes.filter.deps.filter_deletion_service"
+    with patch("src.web.filter.handlers.ChannelAnalyzer") as mock_analyzer, patch(
+        "src.web.filter.handlers.deps.filter_deletion_service"
     ) as mock_svc:
         from src.filters.models import ChannelFilterResult, FilterReport
 
@@ -788,7 +788,7 @@ async def test_purge_selected_with_multiple_pks(route_client, db):
     pk1 = await _add_filtered_channel(db, channel_id=3200, title="Purge 1")
     pk2 = await _add_filtered_channel(db, channel_id=3201, title="Purge 2")
 
-    with patch("src.web.routes.filter.deps.filter_deletion_service") as mock_svc:
+    with patch("src.web.filter.handlers.deps.filter_deletion_service") as mock_svc:
         mock_svc.return_value.purge_channels_by_pks = AsyncMock()
         resp = await route_client.post(
             "/channels/filter/purge-selected",
